+++ tests/test_scrabble_game.py
@@ -1103,6 +1103,334 @@ class TestScrabbleGame(unittest.TestCase):
         scrabbleGame.putWord("araña", (11, 11), "v")
         game.showBoard(scrabbleGame.board)
     
+    def test_validateWord0a100(self):
+        game = Game()
+        scrabbleGame = ScrabbleGame(playerCount=4)
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("R", 8), 
+            Tile("O", 1), 
+            Tile("P", 8), 
+            Tile("E", 3),
+            Tile("N", 1), 
+            Tile("U", 1), 
+            Tile("A", 1)
+        ]
+        scrabbleGame.putWord("pero", (8, 8), "h")
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("L", 8), 
+            Tile("L", 1), 
+            Tile("F", 8), 
+            Tile("D", 3),
+            Tile("M", 1), 
+            Tile("C", 1), 
+            Tile("N", 1)
+        ]
+        scrabbleGame.putWord("del", (7, 9), "v")
+        scrabbleGame.current_player = scrabbleGame.players[2]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("C", 8), 
+            Tile("I", 1), 
+            Tile("N", 8), 
+            Tile("S", 3),
+            Tile("E", 1), 
+            Tile("O", 1), 
+            Tile("R", 1)
+        ]
+        scrabbleGame.putWord("coso", (7, 11), "v")
+        scrabbleGame.current_player = scrabbleGame.players[3]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("N", 8), 
+            Tile("C", 1), 
+            Tile("R", 8), 
+            Tile("U", 3),
+            Tile("S", 1), 
+            Tile("I", 1), 
+            Tile("S", 1)
+        ]
+        scrabbleGame.putWord("no", (10, 10), "h")
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("N", 8), 
+            Tile("U", 1), 
+            Tile("A", 8), 
+            Tile("E", 3),
+            Tile("Y", 1), 
+            Tile("A", 1), 
+            Tile("A", 1)
+        ]
+        scrabbleGame.putWord("suya", (9, 11), "h")
+        scrabbleGame.current_player = scrabbleGame.players[2]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("I", 8), 
+            Tile("N", 1), 
+            Tile("E", 8), 
+            Tile("R", 3),
+            Tile("Z", 1), 
+            Tile("S", 1), 
+            Tile("O", 1)
+        ]
+        scrabbleGame.putWord("sano", (8, 14), "v")
+        scrabbleGame.current_player = scrabbleGame.players[3]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("R", 8), 
+            Tile("S", 1), 
+            Tile("I", 8), 
+            Tile("H", 3),
+            Tile("C", 1), 
+            Tile("C", 1), 
+            Tile("T", 1)
+        ]
+        scrabbleGame.putWord("sanos", (8, 14), "v")
+        scrabbleGame.current_player = scrabbleGame.players[0]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("N", 8), 
+            Tile("E", 1), 
+            Tile("O", 8), 
+            Tile("A", 3),
+            Tile("R", 1), 
+            Tile("A", 1), 
+            Tile("E", 1)
+        ]
+        scrabbleGame.putWord("cenar", (7, 11), "h")
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("I", 8), 
+            Tile("E", 1), 
+            Tile("R", 8), 
+            Tile("Z", 3),
+            Tile("I", 1), 
+            Tile("B", 1), 
+            Tile("E", 1)
+        ]
+        scrabbleGame.putWord("pire", (8, 8), "v")
+        scrabbleGame.current_player = scrabbleGame.players[2]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("R", 8), 
+            Tile("I", 1), 
+            Tile("H", 8), 
+            Tile("C", 3),
+            Tile("C", 1), 
+            Tile("T", 1), 
+            Tile("A", 1)
+        ]
+        scrabbleGame.putWord("cierta", (11, 6), "h")
+        scrabbleGame.current_player = scrabbleGame.players[3]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("O", 8), 
+            Tile("A", 1), 
+            Tile("E", 8), 
+            Tile("T", 3),
+            Tile("O", 1), 
+            Tile("E", 1), 
+            Tile("D", 1)
+        ]
+        scrabbleGame.putWord("toca", (9, 6), "v")
+        scrabbleGame.current_player = scrabbleGame.players[0]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("J", 8), 
+            Tile("D", 1), 
+            Tile("C", 8), 
+            Tile("R", 3),
+            Tile("V", 1), 
+            Tile("G", 1), 
+            Tile("N", 1)
+        ]
+        scrabbleGame.putWord("tocar", (9, 6), "v")
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("Z", 8), 
+            Tile("I", 1), 
+            Tile("B", 8), 
+            Tile("E", 3),
+            Tile("A", 1), 
+            Tile("E", 1), 
+            Tile("T", 1)
+        ]
+        scrabbleGame.putWord("zar", (13, 4), "h")
+        scrabbleGame.current_player = scrabbleGame.players[2]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("E", 8), 
+            Tile("O", 1), 
+            Tile("E", 8), 
+            Tile("D", 3),
+            Tile("C", 1), 
+            Tile("M", 1), 
+            Tile("E", 1)
+        ]
+        scrabbleGame.putWord("remo", (11, 9), "v")
+        scrabbleGame.current_player = scrabbleGame.players[3]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("I", 8), 
+            Tile("B", 1), 
+            Tile("E", 8), 
+            Tile("E", 3),
+            Tile("T", 1), 
+            Tile("S", 1), 
+            Tile("H", 1)
+        ]
+        scrabbleGame.putWord("zares", (13, 4), "h")
+        scrabbleGame.current_player = scrabbleGame.players[0]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("H", 8), 
+            Tile("C", 1), 
+            Tile("O", 8), 
+            Tile("O", 3),
+            Tile("O", 1), 
+            Tile("N", 1), 
+            Tile("L", 1)
+        ]
+        scrabbleGame.putWord("loco", (10, 3), "h")
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("E", 8), 
+            Tile("D", 1), 
+            Tile("C", 8), 
+            Tile("E", 3),
+            Tile("A", 1), 
+            Tile("O", 1), 
+            Tile("A", 1)
+        ]
+        scrabbleGame.putWord("celda", (8, 3), "v")
+        scrabbleGame.current_player = scrabbleGame.players[2]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("I", 8), 
+            Tile("B", 1), 
+            Tile("H", 8), 
+            Tile("P", 3),
+            Tile("I", 1), 
+            Tile("T", 1), 
+            Tile("A", 1)
+        ]
+        scrabbleGame.putWord("bicha", (8, 1), "h")
+        scrabbleGame.current_player = scrabbleGame.players[3]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("H", 8), 
+            Tile("O", 1), 
+            Tile("O", 8), 
+            Tile("N", 3),
+            Tile("R", 1), 
+            Tile("G", 1), 
+            Tile("E", 1)
+        ]
+        scrabbleGame.putWord("robe", (6, 1), "v")
+        scrabbleGame.current_player = scrabbleGame.players[0]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("E", 8), 
+            Tile("O", 1), 
+            Tile("A", 8), 
+            Tile("L", 3),
+            Tile("A", 1), 
+            Tile("H", 1), 
+            Tile("F", 1)
+        ]
+        scrabbleGame.putWord("hola", (5, 5), "v")
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("I", 8), 
+            Tile("T", 1), 
+            Tile("E", 8), 
+            Tile("V", 3),
+            Tile("R", 1), 
+            Tile("P", 1), 
+            Tile("L", 1)
+        ]
+        scrabbleGame.putWord("pero", (6, 2), "h")
+        scrabbleGame.current_player = scrabbleGame.players[2]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("H", 8), 
+            Tile("O", 1), 
+            Tile("N", 8), 
+            Tile("G", 3),
+            Tile("U", 1), 
+            Tile("S", 1), 
+            Tile("A", 1)
+        ]
+        scrabbleGame.putWord("tocarás", (9, 6), "v")
+        scrabbleGame.current_player = scrabbleGame.players[3]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("A", 8), 
+            Tile("C", 1), 
+            Tile("E", 8), 
+            Tile("U", 3),
+            Tile("T", 1), 
+            Tile("S", 1), 
+            Tile("H", 1)
+        ]
+        scrabbleGame.putWord("sueca", (15, 6), "h")
+        scrabbleGame.current_player = scrabbleGame.players[0]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("I", 8), 
+            Tile("T", 1), 
+            Tile("V", 8), 
+            Tile("L", 3),
+            Tile("Q", 1), 
+            Tile("S", 1), 
+            Tile("T", 1)
+        ]
+        scrabbleGame.putWord("suecas", (15, 6), "h")
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("H", 8), 
+            Tile("O", 1), 
+            Tile("N", 8), 
+            Tile("G", 3),
+            Tile("O", 1), 
+            Tile("L", 1), 
+            Tile("X", 1)
+        ]
+        scrabbleGame.putWord("hongo", (5, 5), "h")
+        scrabbleGame.current_player = scrabbleGame.players[2]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("E", 8), 
+            Tile("A", 1), 
+            Tile("A", 8), 
+            Tile("F", 3),
+            Tile("I", 1), 
+            Tile("E", 1), 
+            Tile("I", 1)
+        ]
+        scrabbleGame.putWord("fiera", (3, 4), "v")
+        scrabbleGame.current_player = scrabbleGame.players[3]
+        scrabbleGame.next_turn()
+        scrabbleGame.current_player.tiles = [
+            Tile("A", 8), 
+            Tile("E", 1), 
+            Tile("F", 8), 
+            Tile("D", 3),
+            Tile("Q", 1), 
+            Tile("S", 1), 
+            Tile("H", 1)
+        ]
+        scrabbleGame.putWord("fea", (14, 4), "h")
+        game.showBoard(scrabbleGame.board)
+
     def test_validateWordNotFoundInDictionary(self):
         game = Game()
         scrabbleGame = ScrabbleGame(playerCount=2)